requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
numpy==1.23.5
tensorflow==2.12.0
keras==2.12.0
h5py==3.8.0
pillow==10.0.1
pydantic==1.10.8
opencv-python==4.8.1.78
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
sqlalchemy==2.0.19
python-dotenv==1.0.0
typing-extensions==4.5.0